moved dude logic into own lib
