+++ approaches/optimistic/pom.xml
@@ -106,6 +106,13 @@
 			<scope>system</scope>
 			<systemPath>${basedir}/../shared/Partitioning/target/Partitioning-1.0-jar-with-dependencies.jar</systemPath>
 		</dependency>
+		<dependency>
+			<groupId>der.rdse.hpi.de</groupId>
+			<artifactId>duplicate-detection</artifactId>
+			<version>1.0</version>
+			<scope>system</scope>
+			<systemPath>${basedir}/../shared/duplicate-detection/target/duplicate-detection-1.0-jar-with-dependencies.jar</systemPath>
+		</dependency>
 	</dependencies>
 
 	<build>
