--- approaches/optimistic/src/main/java/de/hpi/ddd/dd/DuplicateDetector.java
@@ -1,15 +1,0 @@
-package de.hpi.ddd.dd;
-
-import java.util.List;
-import java.util.Set;
-
-public interface DuplicateDetector {
-
-    /**
-     * Finds all duplicates within a given block
-     * @param records the records of a block
-     * @return a set of sets representing the identified duplicates grouped in classes
-     */
-    Set<Set<Integer>> findDuplicatesForBlock(List<String[]> records);
-
-}
