+++ approaches/optimistic/src/main/java/test/de/hpi/ddd/similarity/numeric/AbsComparatorTest.java
@@ -1,7 +1,7 @@
 package test.de.hpi.ddd.similarity.numeric;
 
-import de.hpi.ddd.similarity.numeric.AbsComparator;
-import de.hpi.ddd.similarity.numeric.NumberComparator;
+import de.hpi.der.similarity.numeric.AbsComparator;
+import de.hpi.der.similarity.numeric.NumberComparator;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
