+++ approaches/shared/duplicate-detection/src/main/java/de/hpi/der/similarity/numeric/NumberComparator.java
@@ -1,4 +1,4 @@
-package de.hpi.ddd.similarity.numeric;
+package de.hpi.der.similarity.numeric;
 
 public interface NumberComparator {
