+++ approaches/shared/duplicate-detection/src/main/java/de/hpi/der/similarity/string/JaccardIndexComparator.java
@@ -1,4 +1,4 @@
-package de.hpi.ddd.similarity.strings;
+package de.hpi.der.similarity.string;
 
 
 import info.debatty.java.stringsimilarity.Jaccard;
